improve max packet size

+++ src/protocol.h
@@ -4,11 +4,11 @@
 #include "config.h"
 #include <time.h>
 
-#define PACKET_HEAD ((int)(sizeof(int)+sizeof(time_t)))
+#define PACKET_HEAD ((long int)(sizeof(long int)+sizeof(time_t)))
 #define MAX_PACKET_BODY (MAX_PACKET_SIZE - PACKET_HEAD)
 
 struct packet {
-    int length;
+    long int length;
     time_t sendTime;
     char *body;
 };
